Creating io scripts
Starting development of file io system. Will go over with Bryce and Harold how it works so we can structure the .suc and .lnl files. May also restructure game files for github to avoid future problems uploading to repo... I'll get back to that. Couldn't figure out how to get it to read or write ... Will have to work it out on desktop so I can take a better look at it. Pushing to repo