Began adding a UI!
Scripted a couple buttons to swap canvases to show different ui's without actually changing the scene. We can add nice animations and textures to the buttons in the future but for now, I think these will work for educational and testing purposes. I again heavily commented all the code so you can understand whats going on and takes notes if need be. :smiley: