Began testing servers.
uNet, or what unity uses as a built in networking manager, is really stupid. But the way they set it up makes it our only option for a server based game. It's ugly as hell, but it can always be redone. If anyone wants to help then that would be great.